feat: extract transactions component from ckBTCWalletTransactions
